Split ObjectID and ObjectIndex into separate files and add some tests
